requirements.txt
openai>=1.45.0
pandas>=2.2.0
pydantic>=2.6.3
python-dotenv>=1.0.0
pytest>=8.0.0
